// frontend/src/app/shared/interfaces/MovieDetail.ts
export interface MovieDetail {
  movieId: number;
  title: string;
  releaseYear: string;
  description: string;
  thumbnail: string;
  backdrop: string;
  genres: string[];
  duration: number;
  language: string;
  rating: number;
  producer: string;
  director: string;
  casts: string[];
  price: number;
}